corrigindo sinals

+++ learn_lab/signals.py
@@ -16,13 +16,18 @@ def delete_file(instance):
 @receiver(pre_delete, sender=Activity)
 def delete_old_file(signal, instance, *args, **kwargs):
     old_instance = Activity.objects.filter(pk=instance.id).first()
-    delete_file(old_instance)
+
+    if old_instance:
+        delete_file(old_instance)
 
 
 @receiver(pre_save, sender=Activity)
 def update_activity_file(signal, instance, *args, **kwargs):
     old_instance = Activity.objects.filter(pk=instance.id).first()
     is_new_instance = old_instance != instance.file
+
+    if not old_instance:
+        return
 
     if is_new_instance:
         delete_file(old_instance)
